+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+ - Add tasks for Enable, Show and Break tutorial pages
+
 ## [0.2.0] - 2018-12-31
 
 ### Added